+++ backend/src/main/java/com/genesis/unipocket/accountbook/query/persistence/response/AccountBookDetailResponse.java
@@ -2,6 +2,7 @@ package com.genesis.unipocket.accountbook.query.persistence.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.genesis.unipocket.global.common.enums.CountryCode;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -11,7 +12,7 @@ public record AccountBookDetailResponse(
 		String title,
 		CountryCode localCountryCode,
 		CountryCode baseCountryCode,
-		Long budget,
+		BigDecimal budget,
 		@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss")
 				LocalDateTime budgetCreatedAt,
 		List<?> tempExpenseBatchIds,
